back/requirements.txt
#AI
openai==0.27.8
#web framework
Flask==2.2.3
Flask-Cors==3.0.10

python-dotenv